removing ccf_req and replacing with pbft::Request
