+++ CMakeLists.txt
@@ -148,19 +148,6 @@ if(BUILD_TESTS)
       lua.host
       secp256k1.host
     )
-    if (PBFT)
-      if("sgx" IN_LIST TARGET)
-        target_link_libraries(frontend_test PRIVATE
-          -Wl,--allow-multiple-definition #TODO(#important): This is unfortunate
-          libbyz.enclave
-        )
-      else()
-        target_link_libraries(frontend_test PRIVATE
-          -Wl,--allow-multiple-definition #TODO(#important): This is unfortunate
-          libbyz.host
-        )
-      endif()
-    endif()
 
   if(NOT PBFT)
     add_unit_test(membervoting_test
